Dispersion Trading Using Bank Nifty
What is Dispersion?
In Dispersion or Correlation Trading, Trader keeps a view on Volatility. Volatility mean reverts so Dispersion or Correlation also mean reverts.
Long and short both positions are made on volatilities. To deal with directional risk, Positions are delta neutralize by buying or selling future. So directional risk is limited.
Dispersion is traded by selling Index options and buying stock options or Buying index options and Selling stocks options.
Dispersion Strategy Details
Dispersion trading is hedge strategy to take advantage of relative value differences in implied volatilities between index and index components.
Straddle and strangle combination is made by taking short options positions on an index and a long options positions on the components of the index or vice versa.
By keeping individual stock/index options portfolio delta close to zero, the directional risk is saved on each stock and index option portfolio.

Data and Time Frame
The focus of the dispersion trading strategy is Indian Bank Nifty index options and its components bank stocks. We use the 15 minutes data for this strategy. We can also use 30 minutes of data also. It all depends upon, which time frame is giving best results.
Strategy Practical Guideline
Steps taken during the projects are as follow:
Implied volatility calculation
To implement this strategy it is important to first calculate the implied volatility of the index Bank Nifty and the stocks in Bank Nifty.
We use Black-Scholes model to calculate the IV’s. We need time to mature, the strike price, the risk-free interest rate, and the current underlying price
Implied/Dirty correlation calculation
We use implied volatility to calculate the implied correlation between stocks. We divide Index IV by weighted Average stock IV to check IC level.  If implied correlation is high then there is an indication to sell the index options and buy the stock options and vice versa for low implied correlation.
Selection of option
We use a combination of straddle and strangle of both puts and calls. Straddle means ATM call and Puts of the same strike. Strangle means OTM calls and Puts. The initial hedge is made to keep delta close to zero by buying or selling relative stock future.
Continuous Delta Neutral/Dynamic Hedging
This is further hedged using future contracts to keep the whole process delta neutral. Delta of this strategy was adjusted every fifteen minutes. When the delta went above 1, one future contract was sold and when the delta dropped to -1 the delta was neutralized by buying one future contract. It is important to keep the delta close to zero for the duration of the trade.
Example: ICICI Bank

Stoploss and Exit
Along with Entry and continuous Delta Neutral, it is also necessary to have a plan to exit from the dispersion trade when the trade moves in our favour and when the trade moves against our expectation. This risk management protects from unexpected loss or by giving up on the profit earned.
Profit and loss calculation
The PNL for this strategy is coming from two elements one from Options another from future contracts. The future contracts are added to stay delta neutral. At the time of expiry, all positions are squared off and the final profit is calculated.

Dispersion Applications
Dispersion trading is a very profitable low risk hedged strategy and rewards are far greater than Risk. If this strategy is traded automatically after backtest, then more efficient results can be generated. This strategy can be used on any traded index where index future is traded and index component options are liquid enough to make the trades. In Indian or Global markets there are a number of indices available that are suitable for correlation/dispersion trading.